+++ packages/twenty-server/src/workspace/workspace-cleaner/commands/clean-inactive-workspaces.command.ts
@@ -4,7 +4,7 @@ import { Command, CommandRunner, Option } from 'nest-commander';
 
 import { MessageQueue } from 'src/integrations/message-queue/message-queue.constants';
 import { MessageQueueService } from 'src/integrations/message-queue/services/message-queue.service';
-import { CleanInactiveWorkspaceJob } from 'src/workspace/cron/clean-inactive-workspaces/clean-inactive-workspace.job';
+import { CleanInactiveWorkspaceJob } from 'src/workspace/workspace-cleaner/cron/clean-inactive-workspace.job';
 
 export type CleanInactiveWorkspacesCommandOptions = {
   dryRun: boolean;
