+++ packages/twenty-server/src/workspace/workspace-cleaner/commands/start-clean-inactive-workspaces.cron.command.ts
@@ -4,14 +4,14 @@ import { Command, CommandRunner } from 'nest-commander';
 
 import { MessageQueue } from 'src/integrations/message-queue/message-queue.constants';
 import { MessageQueueService } from 'src/integrations/message-queue/services/message-queue.service';
-import { cleanInactiveWorkspaceCronPattern } from 'src/workspace/cron/clean-inactive-workspaces/clean-inactive-workspace.cron.pattern';
-import { CleanInactiveWorkspaceJob } from 'src/workspace/cron/clean-inactive-workspaces/clean-inactive-workspace.job';
+import { cleanInactiveWorkspaceCronPattern } from 'src/workspace/workspace-cleaner/cron/clean-inactive-workspace.cron.pattern';
+import { CleanInactiveWorkspaceJob } from 'src/workspace/workspace-cleaner/cron/clean-inactive-workspace.job';
 
 @Command({
-  name: 'clean-inactive-workspaces:cron:stop',
-  description: 'Stops the clean inactive workspaces cron job',
+  name: 'clean-inactive-workspaces:cron:start',
+  description: 'Starts a cron job to clean inactive workspaces',
 })
-export class StopCleanInactiveWorkspacesCronCommand extends CommandRunner {
+export class StartCleanInactiveWorkspacesCronCommand extends CommandRunner {
   constructor(
     @Inject(MessageQueue.cronQueue)
     private readonly messageQueueService: MessageQueueService,
@@ -20,9 +20,11 @@ export class StopCleanInactiveWorkspacesCronCommand extends CommandRunner {
   }
 
   async run(): Promise<void> {
-    await this.messageQueueService.removeCron(
+    await this.messageQueueService.addCron<undefined>(
       CleanInactiveWorkspaceJob.name,
+      undefined,
       cleanInactiveWorkspaceCronPattern,
+      { retryLimit: 3 },
     );
   }
 }
